roachtest: refactor run/bench code
This commit cleans up some of the clunky roachtest run code. We move
all run code and flags to a separate file. We switch run flags to use
global vars (like all other flags). We also extract out the run/bench
code from the command-wide `PersistentPreRunE`.

Epic: none
Release note: None
