+++ pkg/cmd/roachtest/cluster.go
@@ -104,6 +104,7 @@ var (
 	overrideFlagset  *pflag.FlagSet
 	overrideNumNodes = -1
 	clusterName      string
+	local            bool
 	clusterWipe      bool
 	zonesF           string
 	teamCity         bool
